refactor to use abstract class
