+++ Java/FX-400/FX400.java
@@ -6,29 +6,7 @@ import java.io.File;
 import org.ini4j.Ini;
 
 
-public class FX400 extends Thread{
-
-    protected int DELAY = 200; //Default 200. Delay time for everything. Multiply by delay strength to change length
-    protected double DEVICE_INSERT_DELAY_STRENGTH = 2; // Default 2. Multiplied to DELAY. The Delay time after inserting a device.
-    protected double DEVICE_UPDATE_DELAY_STRENGTH = 1.5; // Default 1.5. Multiplied to DELAY. The Delay time after updating a device (tag name, type, etc).
-    protected boolean BYPASS_PAUSE = false; //Prevents the error prompt from showing
-    protected boolean IGNORE_TAG_LENGTH = false; //Omits tag length requirement from errors
-    protected boolean SKIP_INSERT_DEVICES = false; //Set the bot to update devices only instead of inserting first. Meant for when the devices are entered and other details (type, f1 tags,etc) are untouched.
-    protected String SETTINGS_FILE = "settings.ini";
-
-    protected DataEntryBot bot;
-    protected int skip_count = 19;
-    protected boolean is_running = false; //used to stop the bot from running without closing process
-    protected boolean is_paused = false; //used to prompt user to enable AR related settings -- MAKE USE OF THIS IF PAUSING GUI
-
-    public FX400() {
-        try {
-            readSettings();
-            bot = new DataEntryBot(DELAY);           
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
+public class FX400 extends ConfigBot{
 
     public void run() {
         System.out.println("Starting FX400 Data Entry");
@@ -140,28 +118,7 @@ public class FX400 extends Thread{
         }
     }
 
-    protected void skipDevices() {
-        bot.pressKey(KeyEvent.VK_RIGHT, skip_count);
-    }
-
-    protected void open() {
-        try {
-            Thread.sleep(DELAY);
-            bot.keyPress(KeyEvent.VK_SHIFT);
-            bot.keyPress(KeyEvent.VK_F10);
-            bot.keyRelease(KeyEvent.VK_SHIFT);
-            bot.keyRelease(KeyEvent.VK_F10);
-            bot.delay(DELAY);
-
-            bot.pressKey(KeyEvent.VK_DOWN);
-            bot.pressKey(KeyEvent.VK_ENTER, 1 , DEVICE_INSERT_DELAY_STRENGTH);
-            
-        } catch (Exception e) {
-            System.err.println(e);
-        }
-    }
-    
-    protected void addPhotoDetector() {
+    public void addPhotoDetector() {
         open();
         bot.pressKey(KeyEvent.VK_TAB, 3);
         skipDevices();
@@ -170,7 +127,7 @@ public class FX400 extends Thread{
         bot.pressKey(KeyEvent.VK_END);
     }
 
-    protected void addDuctDetector() {
+    public void addDuctDetector() {
         open();
         bot.pressKey(KeyEvent.VK_P, 5);
         bot.pressKey(KeyEvent.VK_TAB, 2);
@@ -180,7 +137,7 @@ public class FX400 extends Thread{
         bot.pressKey(KeyEvent.VK_END);
     }
 
-    protected void addAlarmInputMod() {
+    public void addAlarmInputMod() {
         open();
         bot.pressKey(KeyEvent.VK_D, 2);
         bot.pressKey(KeyEvent.VK_TAB, 3);
@@ -190,7 +147,7 @@ public class FX400 extends Thread{
         bot.pressKey(KeyEvent.VK_END);
     }
 
-    protected void addNonLatchedSupv() {
+    public void addNonLatchedSupv() {
         open();
         bot.pressKey(KeyEvent.VK_D, 2);
         bot.pressKey(KeyEvent.VK_TAB, 2);
@@ -202,7 +159,7 @@ public class FX400 extends Thread{
         bot.pressKey(KeyEvent.VK_END);
     }
 
-    protected void addLatchedSupv() {
+    public void addLatchedSupv() {
         open();
         bot.pressKey(KeyEvent.VK_D,2);
         bot.pressKey(KeyEvent.VK_TAB,2);
@@ -214,7 +171,7 @@ public class FX400 extends Thread{
         bot.pressKey(KeyEvent.VK_END);
     }
 
-    protected void addHeatDetector() {
+    public void addHeatDetector() {
         open();
         bot.pressKey(KeyEvent.VK_H,3);
         bot.pressKey(KeyEvent.VK_TAB,3);
@@ -224,7 +181,7 @@ public class FX400 extends Thread{
         bot.pressKey(KeyEvent.VK_END);
     }
 
-    protected void addAlarmInputClassA() {
+    public void addAlarmInputClassA() {
         open();
         bot.pressKey(KeyEvent.VK_D, 2);
         bot.pressKey(KeyEvent.VK_TAB, 1);
@@ -236,7 +193,7 @@ public class FX400 extends Thread{
         bot.pressKey(KeyEvent.VK_END);
     }
 
-    protected void addRelay() {
+    public void addRelay() {
         open();
         bot.pressKey(KeyEvent.VK_D);
         bot.pressKey(KeyEvent.VK_TAB, 2);
@@ -246,20 +203,7 @@ public class FX400 extends Thread{
         bot.pressKey(KeyEvent.VK_END);
     }
 
-    protected void updateTags(Zone zone) {
-        try {
-            Thread.sleep(DELAY);
-            bot.pressKey(KeyEvent.VK_ENTER, 1, DEVICE_UPDATE_DELAY_STRENGTH);
-            bot.pasteText(zone.getTag1());
-            bot.pressKey(KeyEvent.VK_ENTER, 1, DEVICE_UPDATE_DELAY_STRENGTH);
-            bot.pasteText(zone.getTag2());
-            bot.pressKey(KeyEvent.VK_ENTER, 1, DEVICE_UPDATE_DELAY_STRENGTH);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    protected void updateType(Zone zone) {
+    public void updateType(Zone zone) {
         try {
             Thread.sleep(DELAY);
             switch(zone.getType()) {
@@ -294,26 +238,7 @@ public class FX400 extends Thread{
         }
     }
 
-    protected void updateRow(Zone zone) {
-        updateTags(zone);
-        updateType(zone);
-        
-        if(zone.isNS()) {
-            bot.pressKey(KeyEvent.VK_N, 1, DEVICE_UPDATE_DELAY_STRENGTH);
-        }
-
-        bot.pressKey(KeyEvent.VK_ENTER, 1 , DEVICE_UPDATE_DELAY_STRENGTH);
-        
-        if(zone.isAR()) {
-            bot.pressKey(KeyEvent.VK_A);
-            bot.pressKey(KeyEvent.VK_ENTER, 1 , DEVICE_UPDATE_DELAY_STRENGTH);
-        }
-
-        bot.pressKey(KeyEvent.VK_ESCAPE);
-        bot.pressKey(KeyEvent.VK_DOWN);
-    }
-
-    protected void updateZone(Zone zone) {
+    public void updateZone(Zone zone) {
         try {
             updateRow(zone);
 
@@ -329,75 +254,8 @@ public class FX400 extends Thread{
         
     }
 
-    protected void enterZoneList(ZoneList zone_list) {
-        try {
-            bot.pressKey(KeyEvent.VK_HOME, 1, 1); 
-            for(Zone zone : zone_list.zones) {
-                System.out.println("Updating: " + zone.getZoneinfo());
-                if(!zone.getType().equals("Blank Device")) {
-                    updateZone(zone);
-                }
-            }
-            
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    protected void readSettings() {
-        try{
-            Ini ini;
-            File ini_file = new File(SETTINGS_FILE);
-
-            //Create settings file if doesn't exist
-            if(!ini_file.exists()) {
-                ini_file.createNewFile();
-            }
-
-            ini = new Ini(ini_file);
-
-            //Add settings if doesn't exist - only checks if ini section exists, not keys
-            if(!ini.containsKey("Key Delay")) {
-                ini.add("Key Delay");
-                ini.putComment("Key Delay", 
-                    String.format("delayTime - Default %s. Delay time for everything. Multiply by delay strength to change duration. \n" +
-                    "#deviceInsertDelayStrength - Default %s. Multiplied to delayTime. The Delay time after inserting a device. \n" +
-                    "#deviceUpdateDelayStrength - Default %s. Multiplied to delayTime. The Delay time after updating a device (tag name, type, etc).",
-                    DELAY, DEVICE_INSERT_DELAY_STRENGTH, DEVICE_UPDATE_DELAY_STRENGTH)
-                );
-                ini.put("Key Delay", "delayTime", DELAY);
-                ini.put("Key Delay", "deviceInsertDelayStrength", DEVICE_INSERT_DELAY_STRENGTH);
-                ini.put("Key Delay", "deviceUpdateDelayStrength", DEVICE_UPDATE_DELAY_STRENGTH);
-            }
-
-            if(!ini.containsKey("Options")) {
-                ini.add("Options");
-                ini.putComment("Options", 
-                    "bypassPause - Prevents the error prompt from showing. \n" +
-                    "#ignoreTagLength - Omits tag length requirement from errors. \n" +
-                    "#skipInsertDevices - Set the bot to update devices only instead of inserting first. Meant for when the devices are entered and other details (type, f1 tags,etc) are untouched."
-                );
-                ini.put("Options", "bypassPause", BYPASS_PAUSE);
-                ini.put("Options", "ignoreTagLength", IGNORE_TAG_LENGTH);
-                ini.put("Options", "skipInsertDevices", SKIP_INSERT_DEVICES);   
-            }
-
-            ini.store(ini_file);
-
-            //Read settings
-            DELAY = ini.get("Key Delay", "delayTime", int.class);
-            DEVICE_INSERT_DELAY_STRENGTH = ini.get("Key Delay", "deviceInsertDelayStrength", double.class);
-            DEVICE_UPDATE_DELAY_STRENGTH = ini.get("Key Delay", "deviceUpdateDelayStrength", double.class);
-            BYPASS_PAUSE = ini.get("Options", "bypassPause", boolean.class);
-            IGNORE_TAG_LENGTH = ini.get("Options", "ignoreTagLength", boolean.class);
-            
-        }catch(Exception e) {
-            e.printStackTrace();
-        }
-    }
-
     //Check each zone to see if it meets the panel's requirements. Returns True if one incorrect device found
-    protected boolean validateZones(ZoneList zone_list) {
+    public boolean validateZones(ZoneList zone_list) {
         boolean invalid_found = false;
         boolean current_zone_valid;
         String zone_errors;
@@ -492,22 +350,5 @@ public class FX400 extends Thread{
             }
         }
         return invalid_found;
-    }
-
-    public void setIsRunning(boolean status) {
-        is_running = status;
-
-        //Set bot to null to prevent further inputs
-        if(!is_running) {
-            bot = null;
-        }
-    }
-
-    public void setIsPaused(boolean status) {
-        is_paused = status;
-    }
-
-    public boolean getIsPaused() {
-        return is_paused;
     }
 }
