+++ Java/FX-400/start.bat
@@ -1,4 +1,4 @@
 python ./tools/excelReader.py
-javac -cp   ../dependencies/* Automation.java FX400.java ZoneList.java Zone.java ../DataEntryBot.java FX2000.java Flexnet.java -d .
+javac -cp   ../dependencies/* Automation.java ConfigBot.java FX400.java ZoneList.java Zone.java ../DataEntryBot.java FX2000.java Flexnet.java -d .
 java --enable-native-access=ALL-UNNAMED -cp "../dependencies/*;." Automation.java
 PAUSE
